Fix client build by moving meal plan span constants off .server.
Route UI was importing MEAL_PLAN_MAX_SPAN_DAYS from a server-only module, which broke react-router production builds.
